+++ src/main/java/com/wllfengshu/security/service/impl/RoleServiceImpl.java
@@ -6,6 +6,7 @@ import com.wllfengshu.security.dao.RoleDao;
 import com.wllfengshu.security.exception.CustomException;
 import com.wllfengshu.security.model.Role;
 import com.wllfengshu.security.service.RoleService;
+import com.wllfengshu.security.utils.AuthUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,11 +23,14 @@ import java.util.Map;
 public class RoleServiceImpl implements RoleService {
 
     @Autowired
+    AuthUtil authUtil;
+    @Autowired
     private RoleDao roleDao;
 
     @Override
     public Map<String, Object> insert(Role entity, String sessionId)throws CustomException {
         log.info("insert entity:{}",entity);
+        authUtil.requiresPermission(sessionId,"insertRole");
         Map<String, Object> result = new HashMap<>();
         roleDao.insert(entity);
         return result;
@@ -35,6 +39,7 @@ public class RoleServiceImpl implements RoleService {
     @Override
     public Map<String, Object> delete(Integer id, String sessionId)throws CustomException {
         log.info("delete id:{}",id);
+        authUtil.requiresPermission(sessionId,"deleteRole");
         Map<String, Object> result = new HashMap<>();
         roleDao.deleteByPrimaryKey(id);
         return result;
@@ -43,6 +48,7 @@ public class RoleServiceImpl implements RoleService {
     @Override
     public Map<String, Object> update(Role entity, String sessionId)throws CustomException {
         log.info("update entity:{}",entity);
+        authUtil.requiresPermission(sessionId,"updateRole");
         Map<String, Object> result = new HashMap<>();
         roleDao.updateByPrimaryKey(entity);
         return result;
@@ -51,6 +57,7 @@ public class RoleServiceImpl implements RoleService {
     @Override
     public Map<String, Object> select(Boolean needPermission,Integer id, String sessionId)throws CustomException {
         log.info("select needPermission:{},id:{}",needPermission,id);
+        authUtil.requiresPermission(sessionId,"selectRole");
         Map<String, Object> result = new HashMap<>();
         if (needPermission){
             Map<String, Object> condition = new HashMap<>();
@@ -64,8 +71,9 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public Map<String, Object> selectAll(Boolean needPermission,Integer pageNo,Integer pageSize,String sessionId)throws CustomException {
-        log.info("selectAll needPermission:{},pageNo:{},pageSize:{},sessionId",needPermission,pageNo,pageSize,sessionId);
+    public Map<String, Object> selectAll(Boolean needPermission,Integer pageNo,Integer pageSize, String sessionId)throws CustomException {
+        log.info("selectAll needPermission:{},pageNo:{},pageSize:{}",needPermission,pageNo,pageSize);
+        authUtil.requiresPermission(sessionId,"selectAllRole");
         Map<String, Object> result = new HashMap<>();
         PageHelper.startPage(pageNo, pageSize);
         PageInfo<Role> pageInfo = null;
